backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
google-generativeai==0.8.3
pydantic==2.10.3
python-dotenv==1.0.1
python-multipart==0.0.12
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic>=0.40.0
google-genai==0.8.0
